+++ subprojects/publish/src/main/java/org/gradle/api/publish/internal/PublishServices.java
@@ -16,6 +16,7 @@
 
 package org.gradle.api.publish.internal;
 
+import org.gradle.api.internal.artifacts.ivyservice.projectmodule.ProjectDependencyPublicationResolver;
 import org.gradle.internal.service.ServiceRegistration;
 import org.gradle.internal.service.scopes.AbstractPluginServiceRegistry;
